feat: Add asset information to chains through denoms_metadata query.
Since information about denom is only obtained through chain info json, it is difficult to respond to changed denom information. So, when a new coin is added in the chain, the exponent is fixed to 6, so it can be applied differently from the actual value. To solve this, query and use denoms_metadata.